# telemetry_apis/docs/swagger.yaml
basePath: /api/v1
definitions:
  main.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  main.TelemetryEntry:
    properties:
      device:
        type: string
      gpu_id:
        type: string
      metric_name:
        type: string
      timestamp:
        type: string
      value:
        type: number
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API to query GPU telemetry data stored in InfluxDB 2.
  title: GPU Telemetry API
  version: "1.0"
paths:
  /gpus:
    get:
      description: Get list of unique GPU IDs from telemetry data
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: List GPUs
      tags:
      - GPUs
  /gpus/{id}/telemetry:
    get:
      description: |-
        Get telemetry data for a specific GPU between start_time and end_time (RFC3339).
        If either timestamp is omitted, defaults to the last 1 hour.
      parameters:
      - description: GPU ID
        in: path
        name: id
        required: true
        type: string
      - description: Start time in RFC3339 format (e.g. 2026-05-30T10:00:00Z)
        in: query
        name: start_time
        type: string
      - description: End time in RFC3339 format (e.g. 2026-05-30T11:00:00Z)
        in: query
        name: end_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.TelemetryEntry'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Get GPU Telemetry
      tags:
      - GPUs
swagger: "2.0"